pyproject.toml
[tool.poetry]
name = "sentiments"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.3"
Flask-WTF = "^1.0.0"
psycopg2-binary = "^2.9.3"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]
black = {version = "^22.1.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
